todo: 	-make rigidbody control rotation instead of navagent 	-transition abilities away from ienumerators 	-make PrefabPool return an instance rather than the actual prefab 	-make PrefabPool hold copies of instantiated prefabs so theres no need to go creating and destroying them all the time 	-generalize range attacks to thrownprojectiles and beam projectiles 	-make jumping and climbing not just teleport directly to the apropriate spot 	-make kunai apply effects when colliding with a hitbox as whel as a physical object
------------------------------------------------------------------------------------------------

fixed bug in UnifiedController where a pathpending status would desynchronise the rb position and the navagent position

added in the concept of a trackedGameObject which is a gameobject whos location and bounds may need to be queried often

added in the concept of NavMeshSurfaceOther which is a navmesh surface that can easily be loaded and unloaded from the game. It is based on the fact that baking navmeshsurfaces can overlap. the external world is a layer and the interior navmesh is its own layer. The key is the interior navmesh bakes a little outside its bounds while the external world navmesh builds right up to the structure. This little bit of overlap is the key. When an agent crosses a trigger threshold, they can be switched from the external world mesh to the internal one by changing the agents walkable layer. this allows the seamless intergration of navmeshsurfaces without the need for navmeshlinks. Known downnsides, no two may overlap, else the trigger calls are ambiguous.
created a custom navmeshsurface implementation heavily based on the origional to implement this

\nOriginal Name: ./180814_FightArena.zip