todo: 	-make rigidbody control rotation instead of navagent 	-transition abilities away from ienumerators 	-make PrefabPool return an instance rather than the actual prefab 	-make PrefabPool hold copies of instantiated prefabs so theres no need to go creating and destroying them all the time 	-generalize range attacks to thrownprojectiles and beam projectiles 	-make jumping and climbing not just teleport directly to the apropriate spot 	-add directional poise to unifiedcontroller 	-make kunai apply effects when colliding with a hitbox as well as a physical object
------------------------------------------------------------------------------------------------
welcome back me! this is why its so important to have a job you love. This current job is taxing on your soul, I havent seen you in a while, bud.

Downloaded and tested a few environments with gaia, bloating the size of each zip, but massively improving look of terrain

dumped AI scripts from fightfightfight in, but did not yet combine with current scripts\nOriginal Name: ./181112_FightArena.zip